+++ src/derivatives_trading_coin_futures/rest_api/apis/mod.rs
@@ -15,8 +15,6 @@ pub mod account_api;
 pub use account_api::*;
 pub mod market_data_api;
 pub use market_data_api::*;
-pub mod portfolio_margin_endpoints_api;
-pub use portfolio_margin_endpoints_api::*;
 pub mod trade_api;
 pub use trade_api::*;
 pub mod user_data_streams_api;